api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
SQLAlchemy==2.0.34
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
pandas==2.2.3
APScheduler==3.10.4
